+++ src/runtime/gecko/l20n-xul.js
@@ -1,28 +1,10 @@
 import { prioritizeLocales } from '../../intl/locale';
-import { contexts, Localization } from '../../bindings/html';
 import { documentReady, getXULResourceLinks } from './util';
-import { valueFromContext } from '../../lib/format';
 import { ResourceBundle } from './resourcebundle';
+import { GeckoLocalization } from './localization';
 
 Components.utils.import('resource://gre/modules/Services.jsm');
 Components.utils.import('resource://gre/modules/IntlMessageContext.jsm');
-
-const functions = {
-  OS: function() {
-    switch (Services.appinfo.OS) {
-      case 'WINNT':
-        return 'win';
-      case 'Linux':
-        return 'lin';
-      case 'Darwin':
-        return 'mac';
-      case 'Android':
-        return 'android';
-      default:
-        return 'other';
-    }
-  }
-};
 
 function requestBundles(requestedLangs = navigator.languages) {
   return documentReady().then(() => {
@@ -40,18 +22,7 @@ function requestBundles(requestedLangs = navigator.languages) {
   });
 }
 
-function createContext(lang) {
-  return new MessageContext(lang, { functions });
-}
-
-document.l10n = new Localization(document, requestBundles, createContext);
-
-document.l10n.interactive.then(bundles => {
-  document.l10n.getValue = function(id, args) {
-    return valueFromContext(
-      contexts.get(bundles[0]), id, args
-    )[0];
-  };
-});
+document.l10n = new GeckoLocalization(document, requestBundles);
 
 window.addEventListener('languagechange', document.l10n);
+Services.obs.addObserver(document.l10n, 'language-update', false);
